First of I reverted to r177 since it had become pretty buggy. Then I added in birdsey's cosinus table mathstuff and some minor code optimizations. Then I added support for more particles, removed some bugs support for more dynamic lights and made those values including frame limit editable in setup.txt
I also fixed the overlay and background effects.

+++ game/Font.c
@@ -212,7 +212,7 @@ void fnt_getTextSize( Font *font, const char *text, int *width, int *height )
  */
 void fnt_drawTextBox( Font *font, const char *text, int x, int y, int width, int height, int spacing )
 {
-  size_t len;
+  int len;
   char *buffer, *line;
 
   if ( !font ) return;
@@ -238,7 +238,7 @@ void fnt_drawTextBox( Font *font, const char *text, int x, int y, int width, int
 void fnt_getTextBoxSize( Font *font, const char *text, int spacing, int *width, int *height )
 {
   char *buffer, *line;
-  size_t len;
+  int len;
   int tmp_w, tmp_h;
 
   if ( !font ) return;
